+++ projects/VS2017.UWP/raylib.App.UWP/App.h
@@ -4,22 +4,46 @@
 
 #include "pch.h"
 
-// Define what header we use for BaseApp.h
-#define PCH "pch.h"
-
-// Enable hold hack
-#define HOLDHACK
-
-#include "BaseApp.h"
-
 namespace raylibUWP
 {
-    ref class App sealed : public BaseApp
+    ref class App sealed : public Windows::ApplicationModel::Core::IFrameworkView
     {
-        public:
+    public:
+        App();
 
-            App();
+        // IFrameworkView methods.
+        void Initialize(Windows::ApplicationModel::Core::CoreApplicationView^ applicationView) override;
+        void SetWindow(Windows::UI::Core::CoreWindow^ window) override;
+        void Load(Platform::String^ entryPoint) override;
+        void Run() override;
+        void Uninitialize() override;
+    private:
+        bool mWindowVisible = true;
+        bool mSuspended = false;
 
-            void Update() override; 
+        void GameLoop();
+        void ProcessGamepads();
+
+        // Helpers
+        int GetRaylibKey(Windows::System::VirtualKey kVey);
+
+        // Events
+        void OnActivated(Windows::ApplicationModel::Core::CoreApplicationView^ applicationView, Windows::ApplicationModel::Activation::IActivatedEventArgs^ args);
+        void OnSuspending(Platform::Object^ sender, Windows::ApplicationModel::SuspendingEventArgs^ args);
+        void OnResuming(Platform::Object^ sender, Platform::Object^ args);
+        void OnWindowSizeChanged(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::WindowSizeChangedEventArgs^ args);
+        void OnVisibilityChanged(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::VisibilityChangedEventArgs^ args);
+        void OnPointerPressed(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::PointerEventArgs^ args);
+        void OnPointerReleased(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::PointerEventArgs^ args);
+        void OnPointerWheelChanged(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::PointerEventArgs^ args);
+        void OnPointerMoved(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::PointerEventArgs^ args);
+        void OnKeyDown(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::KeyEventArgs^ args);
+        void OnKeyUp(Windows::UI::Core::CoreWindow^ sender, Windows::UI::Core::KeyEventArgs^ args);
+    };
+
+    ref class AppSource sealed : Windows::ApplicationModel::Core::IFrameworkViewSource
+    {
+    public:
+        Windows::ApplicationModel::Core::IFrameworkView^ CreateView() override;
     };
 }
