+++ projects/VS2017.UWP/raylib.App.UWP/raylib.App.UWP.vcxproj.filters
@@ -6,7 +6,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="App.h" />
-    <ClInclude Include="BaseApp.h" />
     <ClInclude Include="pch.h" />
   </ItemGroup>
   <ItemGroup>
