+++ projects/VS2017.UWP/raylib.UWP/raylib.UWP.vcxproj.filters
@@ -16,6 +16,7 @@
     <ClInclude Include="..\..\..\src\external\stb_rect_pack.h" />
     <ClInclude Include="..\..\..\src\external\stb_truetype.h" />
     <ClInclude Include="..\..\..\src\external\stb_vorbis.h" />
+    <ClInclude Include="..\..\..\src\uwp_events.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\src\core.c" />
