+++ src/utils.h
@@ -72,76 +72,6 @@ void InitAssetManager(AAssetManager *manager, const char *dataPath);   // Initia
 FILE *android_fopen(const char *fileName, const char *mode);            // Replacement for fopen() -> Read-only!
 #endif
 
-#if defined(PLATFORM_UWP)
-// UWP Implementation Features
-
-// Determine if UWP functions are set and ready for raylib's use.
-bool UWPIsConfigured();
-
-// Function for getting program time.
-typedef double(*UWPQueryTimeFunc)();
-UWPQueryTimeFunc UWPGetQueryTimeFunc(void);
-void UWPSetQueryTimeFunc(UWPQueryTimeFunc func);
-
-// Function for sleeping the current thread
-typedef void (*UWPSleepFunc)(double sleepUntil);
-UWPSleepFunc UWPGetSleepFunc(void);
-void UWPSetSleepFunc(UWPSleepFunc func);
-
-// Function for querying the display size
-typedef void(*UWPDisplaySizeFunc)(int *width, int* height);
-UWPDisplaySizeFunc UWPGetDisplaySizeFunc(void);
-void UWPSetDisplaySizeFunc(UWPDisplaySizeFunc func);
-
-// Functions for mouse cursor control
-typedef void(*UWPMouseFunc)(void);
-UWPMouseFunc UWPGetMouseLockFunc();
-void UWPSetMouseLockFunc(UWPMouseFunc func);
-UWPMouseFunc UWPGetMouseUnlockFunc();
-void UWPSetMouseUnlockFunc(UWPMouseFunc func);
-UWPMouseFunc UWPGetMouseShowFunc();
-void UWPSetMouseShowFunc(UWPMouseFunc func);
-UWPMouseFunc UWPGetMouseHideFunc();
-void UWPSetMouseHideFunc(UWPMouseFunc func);
-
-// Function for setting mouse cursor position.
-typedef void (*UWPMouseSetPosFunc)(int x, int y);
-UWPMouseSetPosFunc UWPGetMouseSetPosFunc();
-void UWPSetMouseSetPosFunc(UWPMouseSetPosFunc func);
-
-// The below functions are implemented in core.c but are placed here so they can be called by user code.
-// This choice is made as platform-specific code is preferred to be kept away from raylib.h
-
-// Call this when a Key is pressed or released.
-void UWPKeyDownEvent(int key, bool down);
-
-// Call when a mouse button state changes
-void UWPMouseButtonEvent(int button, bool down);
-
-// Call when the mouse cursor moves
-void UWPMousePosEvent(double x, double y);
-
-// Call when the mouse wheel moves
-void UWPMouseWheelEvent(int deltaY);
-
-// Call when the window resizes
-void UWPResizeEvent(int width, int height);
-
-// Call when a gamepad is made active
-void UWPActivateGamepadEvent(int gamepad, bool active);
-
-// Call when a gamepad button state changes
-void UWPRegisterGamepadButton(int gamepad, int button, bool down);
-
-// Call when a gamepad axis state changes
-void UWPRegisterGamepadAxis(int gamepad, int axis, float value);
-
-// Set the core window pointer so that we can pass it to EGL.
-void* UWPGetCoreWindowPtr();
-void UWPSetCoreWindowPtr(void* ptr);
-
-#endif      // defined(PLATFORM_UWP)
-
 #ifdef __cplusplus
 }
 #endif
